# client/src/app/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/categories/create": {
      "post": {
        "tags": [
          "categorie"
        ],
        "operationId": "save_9",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/ventes/utilisateur/{idVente}": {
      "get": {
        "tags": [
          "vente"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "idVente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/utilisateurs/utilisateur/{idUtilisateur}": {
      "get": {
        "tags": [
          "utilisateur"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "idUtilisateur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/{idArticle}": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/products/by-code-article/{codeArticle}": {
      "get": {
        "tags": [
          "product"
        ],
        "operationId": "findByCodeArticle",
        "parameters": [
          {
            "name": "codeArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/notification/notification/{id}": {
      "get": {
        "tags": [
          "notification"
        ],
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/fournisseurs/{idFournisseur}": {
      "get": {
        "tags": [
          "fournisseur"
        ],
        "operationId": "findById_4",
        "parameters": [
          {
            "name": "idFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fournisseur"
        ],
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "idFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/commandes-fournisseur/{idCommandeClient}": {
      "get": {
        "tags": [
          "commandeFournisseur"
        ],
        "operationId": "findById_5",
        "parameters": [
          {
            "name": "idCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "commandeFournisseur"
        ],
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "idCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/commandes-client/{idCommandeClient}": {
      "get": {
        "tags": [
          "commandeClient"
        ],
        "operationId": "findById_6",
        "parameters": [
          {
            "name": "idCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "commandeClient"
        ],
        "operationId": "delete_6",
        "parameters": [
          {
            "name": "idCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/clients/client/{idClient}": {
      "get": {
        "tags": [
          "client"
        ],
        "operationId": "findById_7",
        "parameters": [
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/{idCategory}": {
      "get": {
        "tags": [
          "categorie"
        ],
        "operationId": "getCategorieById",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/categories/category/{codeCategory}": {
      "get": {
        "tags": [
          "categorie"
        ],
        "operationId": "findByCodeCategorie",
        "parameters": [
          {
            "name": "codeCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/categories/categories/delete/{idCategory}": {
      "delete": {
        "tags": [
          "categorie"
        ],
        "operationId": "delete_8",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateVente": {
            "type": "string",
            "format": "date-time"
          },
          "commentaire": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string"
          },
          "houseNumber": {
            "type": "integer",
            "format": "int32"
          },
          "zipCode": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "roleName": {
            "type": "string"
          }
        }
      },
      "UtilisateurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "password": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "dateDeNaissance": {
            "type": "string",
            "format": "date"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          }
        }
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codeCategory": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codeProduit": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "prixUnitaireHt": {
            "type": "number"
          },
          "tauxTva": {
            "type": "number"
          },
          "prixUnitaireTtc": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "numberSell": {
            "type": "integer",
            "format": "int32"
          },
          "promo": {
            "type": "boolean"
          },
          "selected": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          }
        }
      },
      "CommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "etatCommande": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "utilisateurDto": {
            "$ref": "#/components/schemas/UtilisateurDto"
          },
          "ligneCommandeClients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LigneCommandeClientDto"
            }
          },
          "livraisonInfoDto": {
            "$ref": "#/components/schemas/LivraisonInfoDto"
          },
          "trackingNumber": {
            "type": "string"
          },
          "paiementDto": {
            "$ref": "#/components/schemas/PaiementDto"
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "LigneCommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productDto": {
            "$ref": "#/components/schemas/ProductDto"
          },
          "commandeClient": {
            "$ref": "#/components/schemas/CommandeClientDto"
          },
          "quantite": {
            "type": "number"
          },
          "prixUnitaire": {
            "type": "number"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LivraisonInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "addressDto": {
            "$ref": "#/components/schemas/AddressDto"
          }
        }
      },
      "PaiementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "montant": {
            "type": "number",
            "format": "double"
          },
          "mode": {
            "type": "string"
          },
          "cardOwner": {
            "type": "string"
          },
          "cardNumber": {
            "type": "number"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "ccv": {
            "type": "integer",
            "format": "int32"
          },
          "statut": {
            "type": "string"
          },
          "commandeClientDto": {
            "$ref": "#/components/schemas/CommandeClientDto"
          }
        }
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "FournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          }
        }
      },
      "CommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "fournisseur": {
            "$ref": "#/components/schemas/FournisseurDto"
          },
          "ligneCommandeFournisseurs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LigneCommandeFournisseurDto"
            }
          },
          "etatCommande": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "LigneCommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productDto": {
            "$ref": "#/components/schemas/ProductDto"
          },
          "commandeFournisseur": {
            "$ref": "#/components/schemas/CommandeFournisseurDto"
          },
          "quantite": {
            "type": "number"
          },
          "prixUnitaire": {
            "type": "number"
          }
        }
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "password": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/AddressDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}